prj_2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.35.0
pycocotools
pyyaml
Pillow
numpy
tqdm
